[pyspark] factor out pyspark to support typing
Currently, pyspark will run into circular imports issue when enabling
typing for model.py. So this PR tried to refactor out pyspark a little
bit to avoid this.
